Use new libpyramicio library which uses MILLISECONDS instead of SECONDS as function call parameters. This allows finer granularity for time-related operations.
Fix error in audio buffer size. The parameter "durationInMilliseconds" is for the complete audio buffer, not for half the audio buffer. Therefore, we were obtaining 2 times less audio than what we expect in the EASY_DSP main daemon. The main daemon has been fixed to account for this detail.

Modified variable used to store return value of recv() and write() to ssize_t instead of int, as per their API interface.

Profiled the system and determined the time it takes to send data to a client. An upper bound is now available in browser-config.h with explanations of why such a bound is needed. Note that the bound is computed, but not applied in the code yet. The code still needs to modified to reject new connections if the maximum number of clients is already reached.