Fix endianess issue with alternative vectorized implementation.
The buffer build-up at the server was due to the client not processing the incoming buffers fast enough. It used to take 570 ms to process one buffer, therefore causing TCP to push back at the server and stop sending data so fast.

The new implementation takes advantage of vectorization to speed up the packet processing. It now takes 20 ms to process one packet, which is a 28x speedup compared to Basile's original version.